Done some refactorings. Created separate class for external command exception.
