+++ hw01/Architecture.png
