+++ hw01/PrescriptiveArchitecture.png
